# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TalentMesh API",
    "version": "0.1.0",
    "description": "Generated from app/api/ by scripts/generate-openapi.ts — do not edit by hand, run `npm run docs:openapi`.\n\nAuth is a session cookie (`tm_access_token`), not a bearer header. Roles and permissions shown as `x-allowed-roles` / `x-required-permission` are read from the withApi() options that actually enforce them.\n\nOperations flagged `x-undocumented-body` do not use withApi(), so their request shape is not declared anywhere machine-readable and has been left blank rather than guessed."
  },
  "servers": [
    {
      "url": "https://talentmeshsolutions.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_access_token"
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/jobs/{jobId}": {
      "patch": {
        "operationId": "patch_api_jobs_jobId",
        "tags": [
          "jobs"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-allowed-roles": [
          "recruiter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "minLength": 50,
                    "maxLength": 12000
                  },
                  "requirements": {
                    "anyOf": [
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "skills_required": {
                    "anyOf": [
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "full-time",
                      "part-time",
                      "contract",
                      "freelance",
                      "internship",
                      "remote",
                      "hybrid"
                    ]
                  },
                  "location": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "salary_min": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "salary_max": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "INR",
                      "USD"
                    ]
                  },
                  "experience_min": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 30
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "experience_max": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 30
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "department": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 80
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "x-body-schema": "jobUpdateSchema",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "403": {
            "description": "Forbidden — role or permission check failed"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/newsletter/resend-confirmation": {
      "post": {
        "operationId": "post_api_newsletter_resend-confirmation",
        "tags": [
          "newsletter"
        ],
        "parameters": [],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-undocumented-body": "Route does not use withApi(); request shape is not declared in a Zod schema and was not inferred.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/newsletter/subscribe": {
      "post": {
        "operationId": "post_api_newsletter_subscribe",
        "tags": [
          "newsletter"
        ],
        "parameters": [],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-undocumented-body": "Route does not use withApi(); request shape is not declared in a Zod schema and was not inferred.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}